fix(#40): 중복 선언된 컨트롤러 제거 및 잘못된 변수 명 수정
